Refactor: 팔로워 및 팔로잉 컴포넌트를 팔로우 컴포넌트로 통합 한 컴포넌트에서 관리할 수 있도록 개선 Related to: #115
